+++ tests/test_stdin.cpp
@@ -44,7 +44,7 @@ static std::pair<std::string, int> runRyStdin(const std::string &code) {
     std::array<char, 4096> buf;
     ssize_t n;
     while ((n = read(pipeOut[0], buf.data(), buf.size())) > 0) {
-        output.append(buf.data(), n);
+        output.append(buf.data(), static_cast<size_t>(n));
     }
     close(pipeOut[0]);
 
@@ -120,7 +120,7 @@ TEST(StdinExecution, FileExecutionStillWorks) {
     std::array<char, 4096> buf;
     ssize_t n;
     while ((n = read(pipeOut[0], buf.data(), buf.size())) > 0) {
-        output.append(buf.data(), n);
+        output.append(buf.data(), static_cast<size_t>(n));
     }
     close(pipeOut[0]);
